reafactor(chatbar): move to separate file
